fix header

+++ be/src/util/disk_info.h
@@ -29,6 +29,8 @@
 
 namespace doris {
 
+#include "common/compile_check_begin.h"
+
 // DiskInfo is an interface to query for the disk information at runtime.  This
 // contains information about the system as well as the specific data node
 // configuration.
@@ -42,7 +44,7 @@ public:
     // Returns the number of (logical) disks on the system
     static int num_disks() {
         DCHECK(_s_initialized);
-        return _s_disks.size();
+        return static_cast<int>(_s_disks.size());
     }
 
     // Returns the 0-based disk index for 'path' (path must be a FS path, not
@@ -101,5 +103,7 @@ private:
 
     static void get_device_names();
 };
+
+#include "common/compile_check_end.h"
 
 } // namespace doris
